requirements.txt
crawl4ai
aiohttp
streamlit
nest-asyncio
pypdf
# Optional for interactive crawling
playwright

# Core
pydantic-settings
aiosqlite

# OCR Pipeline
pytesseract
pdf2image
pyspellchecker
pillow
